apis/feature_extraction/inference/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0                # via uvicorn
fastapi==0.52.0           # via -r requirements.in
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
numpy==1.18.1             # via -r requirements.in
pillow==7.0.0             # via -r requirements.in
pydantic==1.4             # via fastapi
starlette==0.13.2         # via fastapi
torch==1.4.0              # via -r requirements.in
uvicorn==0.11.3           # via -r requirements.in
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
